Convert to ubyte8 prior to image_as_uint call

+++ AxonDeepSeg/ads_utils.py
@@ -279,8 +279,7 @@ def imread(filename, bitdepth=8):
         raw_img = imageio.imread(filename)
         if len(raw_img.shape) > 2:
             raw_img = imageio.imread(filename, as_gray=True)
-
-    img = imageio.core.image_as_uint(raw_img, bitdepth=bitdepth)
+    img = imageio.core.image_as_uint(raw_img.astype(np.uint8), bitdepth=bitdepth)
     return img
 
 def imwrite(filename, img, format='png'):
